day07_data_spark/requirements.txt
pyspark>=3.5.0
pyyaml>=6.0
pandas>=2.2.0
numpy>=1.26.0